Instantiating many times the same template ...
The term 'instantiating many times the same template' is awkward English usage. The best way I can explain is that the 'many times' bit refers to the expression 'instantiating the same template' and breaks up the the flow.  While I was at it, I modified the rest of the sentence a bit.